+++ examples/text_emoji_svg.py
@@ -10,6 +10,10 @@ from lottie import objects
 from lottie import Color, Point
 from lottie.utils.font import FontStyle
 
+parser = script.get_parser()
+parser.add_argument("text", nargs="?", default="Hello\nWorld\nF\U0001F600O\nE🇪🇺U")
+parser.add_argument("--emoji", default="twemoji/assets/svg/")
+ns = parser.parse_args()
 
 an = objects.Animation(120)
 layer = objects.ShapeLayer()
@@ -18,11 +22,11 @@ an.add_layer(layer)
 # The font name "Ubuntu" here, can be detected among the system fonts if fontconfig is available
 # Otherwise you can use the full path to the font file
 # `emoji_svg` needs to point to a directory with the supported emoji as svg
-style = FontStyle("Ubuntu", 128, emoji_svg="twemoji/assets/svg/")
-t = layer.add_shape(style.render("Hello\nWorld\nF\U0001F600O\nE🇪🇺U"))
+style = FontStyle("Ubuntu", 128, emoji_svg=ns.emoji)
+t = layer.add_shape(style.render(ns.text))
 t.transform.position.value.y += t.line_height
 layer.add_shape(objects.Fill(Color(0, 0, 0)))
 layer.add_shape(objects.Stroke(Color(1, 1, 1), 2))
 
-script.script_main(an)
+script.run(an, ns)